Separation eMail base and resolve PR requests
